refactor: change in bounds builder, adjust closing spec in instagram preset, depricate bounds builder method in favor of obj config.
