+++ examples/expo-router-example/app/examples/apple-music/(tabs)/(home,new,radio,library,search)/index.tsx
@@ -1,13 +1,20 @@
 import { router } from "expo-router";
-import { Button, Text, View } from "react-native";
+import { View } from "react-native";
+import Transition from "react-native-screen-transitions";
 
 export default function IndexScreen() {
 	return (
-		<View>
-			<Text>Index Screen</Text>
-			<Button
-				title="Go to Id"
+		<View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
+			<Transition.Pressable
+				sharedBoundTag="test"
 				onPress={() => router.push("/examples/apple-music/123")}
+				style={{
+					width: 200,
+					height: 300,
+					backgroundColor: "red",
+					justifyContent: "center",
+					alignItems: "center",
+				}}
 			/>
 		</View>
 	);
