+++ examples/expo-router-example/app/examples/apple-music/(tabs)/_layout.tsx
@@ -6,7 +6,7 @@ const ICON_SIZE = 26;
 
 export default function TabLayout() {
 	return (
-		<Tabs screenOptions={{ tabBarActiveTintColor: "red" }}>
+		<Tabs screenOptions={{ tabBarActiveTintColor: "red", headerShown: false }}>
 			<Tabs.Screen
 				name="(home)"
 				options={{
