+++ packages/react-native-screen-transitions/src/types/bounds.ts
@@ -1,4 +1,5 @@
 import type { MeasuredDimensions, StyleProps } from "react-native-reanimated";
+import type { BoundsBuilderOptions } from "../utils/bounds/_types/builder";
 import type { ScreenPhase } from "./core";
 
 /**
@@ -77,6 +78,8 @@ export type BoundEntry = {
 	styles: StyleProps;
 };
 
-export type BoundsAccessor = ((id?: string) => BoundsBuilder) & {
+export type BoundsAccessor = ((
+	id?: string | BoundsBuilderOptions,
+) => BoundsBuilder | StyleProps) & {
 	get: (id?: string, phase?: ScreenPhase) => BoundEntry;
 };
